Enable interrupts after scheduling system is online
Prevents a possible race condition where we'll hit an interrupt before we
insert the kinit task causing the scheduler to select kidle as the next
task. However, the init() function expects the next task switch to save the
context into kinit so we can continue initialization, but that'll never
happen because we're stuck in kidle forever.