initfs: Clean up pread and pwrite
Trim pread calls if they try to read outside of the end of the buffer so we
don't accidentally cause a segfault.

pwrite currently always reallocates the entire buffer, but that is incredibly
inefficient. This commit updates pwrite to reallocate if the write goes beyond
the end of the buffer. Otherwise, writes will not cause a reallocation.