mm: Improve locate_range algorithm
Instead of always calculating distance from the beginning of a zone, we know
also consider distance from anywhere in the middle if the hint indicates a
location above the zone.

For example, if there are free ranges from 0 - 100, 200 - 300, and the hint
wants 120 with size 20, the old algorithm would have returned 200. However,
80 is clearly closer with a distance of 40, and this algorithm now takes
that into consideration.