x86_64: Fix page flag corruption
The page mapping code used __set_flags without filtering out the nonsensical
flags for the PML4, PDPT, and PD entries. This resulted in flags such as
NX being applied to all the pages in that particular section. For example,
if PAGE_NX was passed in to map a userspace page, all the entries in that
PDPT slot would have NX applied to it. This manifested in page faults that
complained about NX during instruction fetches but examining the page itself
showed that the NX bit was not set.

This is fixed by explicitly setting only PAGE_WRITABLE and PAGE_USER (if it
is passed in as a flag) for page directories and using flags only for the
actual page table entry itself.

In the future we should probably rework this so that future mappings don't
overwrite previous mappings' flags.

This also is a step towards fixing #2, although the commit to actually use
NX where appropriate is still missing.