elf: Map with PAGE_WRITABLE then unset after copying
Now that WP is enforced for Ring 0 we need to be careful about ensuring that
pages we write to in kernel mode are writable.

The ELF loader needs write access to load the binary data, so we initially map
the segment with PAGE_WRITABLE. Once we finish loading the binary, we can
remove the write bit.