elf: Copy only p_filesz and zero until p_memsz
We previously copied data for each segment until p_memsz. This worked, but
in reality was copying garbage after p_filesz due to the fact that the real ELF
binary only contains data up to p_filesz. We luckily did not crash because we
located the BSS section and zeroed that, and p_memsz never crossed the boundary
of the initrd mapping.

Per the ELF specification, the file only goes up to p_filesz, and if p_memsz
is > p_filesz we need to zero the extra memory in the program space.

This commit updates the behavior so we only copy p_filesz bytes from the
program and then memset the remaining difference between p_memsz and p_filesz.