x86_64: Implement faster memcpy
This implementation is based on the Linux kernel implementation.

It uses rep movsb and rep movsq to first move 8 bytes at a time, then
copies the remaining individual bytes one-by-one. This in theory is one of
the faster implementations that work on nearly all computers, but of course
it can be optimized more aggressively.

TODO: Check for ERMS and get rid of rep movsq in that case