time: Disable interrupts during time update
Timer interrupts could occur in the middle of an existing update() call,
causing a condition where once the interrupt exists the first update()
call will still be using an old counter value. This resulted in underflows
if directly adding, since the time delta will then be negative.

The initial idea was to use locks, but using locks wouldn't solve the problem
and would instead lead to a deadlock. If update() is called normally and
obtains the lock, and then an interrupt occurs that tries to grab the lock(),
it'll wait forever because interrupts can't be interrupted.

Instead, we disable interrupts during the update() and restore whatever the
setting was before once we're done. We can't just enable interrupts after
because doing so if update() is called from an interrupt will cause
nested interrupts.