log: printk: Use %llu instead of %lu
On 64-bit platforms %lu has the same effect as %llu, meaning that it gets an
entire 64-bit integer from va-args.

However, on 32-bit platforms %lu behaves slightly different, usually having
a size of 32 bits. This is a problem if we're passing in 64-bit integers
to %lu because it will cause corruption in the output.