Add %z format specifier
It turns out that it wasn't a bug when we got garbage if we had a type
specifier mismatch. The C standard says that it's undefined behavior, so we
are free to do whatever.

The standard solution is the %z specifier, which expands to the size of
size_t. Then, whereever we have size_t, we use %z which then ensures that
we have the correct size for va_arg.