proc: elf: Unify architectures
No need to have each architecture manually load ELF. This means less
maintainence burden as we only need to update one file instead of
multiple ones.

To support this, architectures are now required to have a macro called
WORD_SIZE in their types.h that specifies the size of a word for ELF
files.