x86_64: syscall: Simulate an interrupt frame
Before, one of the issues was that system calls could not modify
the system call return state. While most system calls did not
need this, some calls such as signal handlers and execve did.

Therefore, the syscall handler now builds an interrupt frame
exactly like an interrupt handler, passes it in to the system
call handler so system calls can access and modify it, and
then restores the state.