sched: Remove noreturn attribute from yield()
yield() was originally marked with the noreturn attribute because at
first glance, it doesn't appear to return. Instead, it would transfer
control to the next process.

However, as it turns it, there are issues with the way that yield()
is implemented. yield() uses interrupts, which return to the address
following the int instruction. Once the process is scheduled again,
control will go to the nop instruction and then fall through into
the next function, which happened to be setup_signal. In essence,
it meant that yield() did exactly as it's name implies, but fails
to return control back to the calling code.

This was never noticed because yield() was only ever called once a
thread was descheduled, meaning that it will never run again.
Therefore, the fall through into the setup_signal code never mattered
since it didn't even run.

However, when implementing wait support, yield was needed, and it
started manifesting as page faults in memcpy since setup_signal
uses memcpy. Since we never called setup_signal intentionally,
it passed garbage parameters into memcpy which then wrote into
random memory.

This commit removes the noreturn attribute so GCC compiles code with
a ret instead of nop so that the code correctly returns to the code
that called yield() once the calling thread is scheduled again.