lib: list: Add iterator_to and get rid of remove()
iterator_to returns an iterator given a value. This is useful if we are
planning to erase it later which makes removal O(1) now.

Previously removal was O(n) since we had to loop through the list
looking for the correct iterator, but now we can directly create it.

Of course, using iterator_to on a non-member is undefined behavior, but
will most likely crash.

Now, we can get rid of remove() since remove() basically did what a
call to iterator_to then erase did