proc: sched: Remove waiter from list as soon as it's woken
Previously we have the waiter remove itself in wait(), but this started
causing problems if multiple interrupts were queued causing wakeup()
to be called multiple times.

What would happen is that the thread would be inserted into the run
queue multiple times, but since we're using intrusive linked lists,
this started causing corruption of the run queue and eventual crashes
since the run queue would become empty.

Instead, we have the thread immediately removed in wakeup(), which
prevents it from being inserted multiple times by consecutive calls to
wakeup().

For signals, since they are not routed through wakeup(), we explicitly
check if we were woken up by a signal when returning from wait() and
remove ourselves from the queue.