x86: Include bootstrap section in kernel zone
Originally the bootstrap section was not included in the kernel zone
that the physical memory allocator refused to allocate from because the
idea was that we could free up that memory after entering C++ code.

However, as it turns out, the page tables are stored there, and
overwriting them would cause the page table to be nonsense.

This was never caught before because the previous stack allocator gave
out addresses near the end of the available memory since they were the
last ones added. However, the new buddy allocator cares less and gave
out the address that the page table was at, causing kernel crashes.

This commit moves kernel_start to the beginning of the linker script,
meaning that the entire kernel is encompassed.