Use crt* files and call global constructors correctly
We used to have a special linker section with symbols at both the start
and end to store pointers to constructors. Upon startup, before we
started running any serious C++ code, we run through that section and
call the functions pointed to by the pointers in that section to call
the constructors for all global objects.

However, this method seems to be deprecated in the osdev.org wiki,
where we initially got the implementation for this from. Furthermore,
there are some other stuff that GCC provides in the startup files that
we would get if we were in userspace but we disabled using -nostdlib
such as setting up a dummy frame.

Additionally, the official GCC documentation located at
https://gcc.gnu.org/onlinedocs/gccint/Initialization.html specify that
there are only two valid variants for when we can define arbitrary
sections (which we can). One, is to call the GCC provided _init()
function. The other is that GCC will automatically inject a call to the
initialization function before main(), but we don't have a main().

Accordingly, this commit updates the build system to use crt* files,
such as crti and crtn (provided by architecture) and crtbegin and
crtend (provided by GCC). Then, we can initialize everything by
calling the provided _init() function in asm_to_cxx_trampoline.

One downside of this is that we are somewhat hacking CMake to do
what we want since GCC is very picky about the order of linking those
crt* files. The worst offender is hardcoding the linking command
string, which is indeed liable to break when CMake changes the internal
format of that.