x86: Fix multiboot end calculation to avoid clobbering it
Instead of using multiboot_start which is rounded down, we need to use
the raw pointer as the start. This is because it's possible that
the difference between the raw pointer and rounded value is large
enough to cause the round_up macro to not round to the correct value.

For example: Assume the start is at 0x600 and the size is 0x500. Then,
multiboot_start will be at 0. Since the calculation is start + size,
using the multiboot_start as the base will be 0x1000, while the raw
pointer will round to 0x2000, the correct value.