fs: block: Replace read/write with request()
The block layer itself still presents a read/write interface to the
filesystem, but block drivers now only need to implement a request()
function that accepts a BlockRequest pointer.

This makes drivers much easier to write, as the block layer now handles
a lot of things. For example, the block layer now builds a sglist for
the driver to directly use based on the maximum sg region size and
count, handles copying the DMA buffer data to the userspace buffer,
calculates # of blocks, and also the start block address.

This interface will likely change in the future for block drivers.
Currently, a request() function is called synchronously with a single
block request at a time. However, many drivers, AHCI included, are
capable of handling multiple requests. Therefore, in the future, we
may move the request handling into the driver itself (like how Linux)
so it can decide how many requests to handle at a time.