fs: initrd: Check for magic string
tar headers contain the magic string "ustar\0". We used to check if the
magic was null, but this commit adds checking for the real magic string.

This should in theory at least guarantee that we're dealing with a real
tar file, but it doesn't guarantee the validity of the data since we're
not checking file checksums yet.