mm: buddy: Use std::vector<bool> instead of bitset
std::vector<bool> has a specialization for bool where it uses bits
instead of entire bools. This basically means that it is a bitset.

Unfortunately, we couldn't use std::bitset since it requires us to know
the size of the bitset at compile time which of course isn't possible
since we don't know how much memory is available yet.