Revert "CMakeLists.txt: Enable LTO"
Unfortunately this breaks modules since LTO assumes that certain
functions can either be completely removed or, more likely, that they
can be inlined which is a valid assumption if considering the kernel
itself only.

However, modules may (and often do) use those removed or inlined
symbols, but since LTO removed them the module loader will be unable
to resolve the missing symbols.

The Linux kernel had a patch for LTO that basically added an attribute
to the export_symbol() macro that made LTO ignore that function.
Similarly, the missing symbols in Quark can be resolved by adding the
attributes noinline,used.

However, this is really tedious, and the benefits of LTO are not that
signficant. Thus, let's just remove LTO for now. The compiler
infrastructure is there, so really this just adjusts the build files
to disable it.

This reverts commit 4d5d3c89730783c4c8ff69bc9c30d9f097976ce4.