kernel: module: Save pointer to module structure before erasing
Before we would erase the iterator then use it as a pointer to the
module structure which is a big no-no since erasing invalidates
the iterator. That meant that we were in reality deleting random
stuff.

This would manifest in really strange bugs after unloading a module.

This commit saves the pointer before erasing and uses the saved pointer
as the argument to delete.