mm: physical: Switch to stack allocator
Buddy allocator was broken because it didn't free pages properly when
coalescing, leading to significant memory leaks.

To fix this, we would need to somehow free an arbitrary page from a
stack which is virtually impossible without resorting to a linked
list.

To implement a linked list of pages, we would need to a) map the entire
physical address space into memory which is undesirable for several
reasons or b) change the mapping of a specific page to access each
physical page (similar to how Linux implements high memory).

I tested b), which had significant performance penalties as expected
for the numerous TLB flushes.