proc: Make idle thread run only if no more threads are runnable
Instead of making the idle thread a regular thread in the runqueue,
the idle thread should only run if no more threads are runnable.

This prevents the idle thread from using CPU time when there are other
threads that need it. For example, if there is a single thread, the
CPU will be split 50/50 between the idle thread and the other thread
under the old scheme. After this commit, the other thread will receive
100% of the CPU time.