x86_64: mm: Remove direct RAM map
Instead of 512GB, in theory x86_64 supports a theoretical maximum of
512 GB / sizeof(struct page) * 4 KiB = 128 TB of RAM now.

Of course, if struct page increases in size this will go down, so
priority is on keeping struct page as minimal as possible.

Another option is giving pagedb access to more RAM. For example, giving
it two PML4 slots would give it 1 TB which would double the amount of
RAM that Quark supports.