x86_64: mm: Free as large chunks of memory as possible at once
Previously we freed one page of memory (4096 bytes) at a time when
initializing the buddy allocator. This was somewhat inefficient as the
buddy allocator needed to spend time coalescing the blocks into larger
blocks.

This commit introduces an algorithm to free the largest possible chunk
of memory as possible in one pass, subject to memory type limitations.

With this commit, the initialization process is significantly faster,
nearly instantaneous when built in release mode.