x86_64: Setup a very tiny stack for bootstrap32
Turns out Multiboot 2 leaves ESP undefined, in the past we didn't run
into any problems using whatever GRUB setup, but it seems to have
broken recently.

Since we use pushfd to check for the existence of CPUID, setup a stack
so that pushfd will work. It's very small because we're not using it
for anything else, and bootstrap64 will setup a much larger stack.