proc: thread: Make tid and parent private
Making tid private requires reversing the order of thread creation.
Previously, the thread was first created, then a TID was assigned
when it registered with the parent process. However, since tid is
now private, the parent process can't set the TID anymore.

Instead, now the parent process will first determine a TID, then
create the thread passing in a pointer to itself and the TID which
the thread will then load.