proc: thread: Refresh signal status on creation
Since we don't initialize signal_required explicitly, often times
it'll be set to true due to whatever memory we're using. Under most
conditions this wasn't noticed due to the fact that the list of
pending signals was properly initialized to empty, so no signal
would be delivered.

However, when implementing the upcoming changes to the wait queue
system which checks for a pending signal using is_signal_pending()
and aborts the wait if there is a signal, this manifested as a
premature exit from the wait which was not intended.